+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 		- `CancelAndDispose(CancellationTokenSource)`
 		- `FindInAllChildren(Transform)`
 - `GetComponentCached` and `TryGetComponentCached` now also extend `Component` instead of just `GameObject`
+- `ScriptableList<TItem>`: A quick way to define a `ScriptableObject` that holds a list of items and implements `IList`
 
 ### Fixed